CP - fixup: testlarge: implement -max-step option
This commit corrects an erroneous choice of type for a temporary in
the logic implementing the `-max-step` option in `testlarge.c`.  Prior
to this change, it was not possible to test payloads larger than 2GiB
(a regression relative to before this option was added).

+++ tests/testlarge.c
@@ -112,7 +112,7 @@ void _print_stat(int myproc, stat_struct_t *st, const char *name, int operation)
 // Double payload at each iter, subject to max_step
 // but include max_payload which may not otherwise be visited
 #define ADVANCE(sz) do {                           \
-        int step = MIN(max_step, sz);              \
+        size_t step = MIN(max_step, sz);           \
         if (!sz) {                                 \
           sz = 1;                                  \
         } else if (sz < max_payload && sz+step > max_payload) { \
